:recycle: [refactor] useUser 리팩토링
